# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skolva API
  version: 0.2.0
  description: >
    Skolva — self-hosted community management platform.
    Phase 2: Identity (auth, users, roles, permissions).
servers:
  - url: /api
    description: API root
paths:
...
  /auth/login:
    post:
      summary: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/register:
    post:
      summary: Create a user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/2fa/setup:
    post:
      summary: Start 2FA setup (returns provisioning URI + recovery codes)
      operationId: setup2FA
      responses:
        "200":
          description: Provisioning URI and recovery codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setup2FAResponse"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/2fa/confirm:
    post:
      summary: Confirm TOTP code to activate 2FA
      operationId: confirm2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
      responses:
        "204":
          description: 2FA activated
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/2fa/verify:
    post:
      summary: Verify 2FA code to complete login
      operationId: verify2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FARequest"
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/2fa/recovery:
    post:
      summary: Use a recovery code to complete login
      operationId: recover2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FARequest"
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/2fa/disable:
    post:
      summary: Disable 2FA
      operationId: disable2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
      responses:
        "204":
          description: 2FA disabled
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /auth/password/forgot:
    post:
      summary: Request a password reset link
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordForgotRequest"
      responses:
        "200":
          description: Reset link sent (if email exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "422":
          $ref: "#/components/responses/Error"
  /auth/password/reset:
    post:
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Password updated
        "422":
          $ref: "#/components/responses/Error"
  /users:
    get:
      summary: List users
      operationId: listUsers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      summary: Get a user
      operationId: getUser
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update a user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete a user
      operationId: deleteUser
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /search/users:
    get:
      summary: Search users (full-text)
      operationId: searchUsers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /users/{id}/roles:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      summary: List a user's roles
      operationId: listUserRoles
      responses:
        "200":
          description: The user's roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      summary: Assign a role to a user
      operationId: assignRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleRequest"
      responses:
        "200":
          description: The user's updated roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /users/{id}/roles/{slug}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - name: slug
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Remove a role from a user
      operationId: removeRole
      responses:
        "204":
          description: Role removed
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Role:
      type: object
      properties:
        slug:
          type: string
        display_name:
          type: string
        is_protected:
          type: boolean
    AssignRoleRequest:
      type: object
      required:
        - role_slug
      properties:
        role_slug:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        is_active:
          type: boolean
        is_protected:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        email:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
    UpdateUserRequest:
      type: object
      required:
        - first_name
        - last_name
      properties:
        first_name:
          type: string
        last_name:
          type: string
        is_active:
          type: boolean
    Setup2FAResponse:
      type: object
      properties:
        provisioning_uri:
          type: string
        recovery_codes:
          type: array
          items:
            type: string
    CodeRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
    Verify2FARequest:
      type: object
      required:
        - temp_token
        - code
      properties:
        temp_token:
          type: string
        code:
          type: string
    PasswordForgotRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
    PasswordResetRequest:
      type: object
      required:
        - user_id
        - token
        - password
      properties:
        user_id:
          type: string
          format: uuid
        token:
          type: string
        password:
          type: string
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"